Move replacement logic to Language.from_config
